Use Display trait to format ExitStatus
Prior to this change, ExitStatus would print errors to stderr using
`"{:?}"` as the format string.  This allowed no flexibility to choose
between `"{:?}"` and `"{}"`.

This change introduces DebugExitStatus and changes ExitStatus to use
`"{}"` whereas DebugExitStatus will use the old behaviour of ExitStatus.
The default type parameter is also changed to maintain the old
behaviour.